+++ crates/entrypoint/src/cli/command/azure_devops/repo/list/azure_devops_repo_list_cli.rs
@@ -1,6 +1,7 @@
 use clap::Args;
 use cloud_terrastodon_azure_devops::AzureDevOpsProjectArgument;
 use cloud_terrastodon_azure_devops::fetch_all_azure_devops_projects;
+use cloud_terrastodon_azure_devops::fetch_azure_devops_repos_batch;
 use cloud_terrastodon_azure_devops::fetch_all_azure_devops_repos_for_project;
 use cloud_terrastodon_azure_devops::get_default_organization_url;
 use eyre::Result;
@@ -10,25 +11,37 @@ use std::io::stdout;
 /// List Azure DevOps repositories in a project.
 #[derive(Args, Debug, Clone)]
 pub struct AzureDevOpsRepoListArgs {
-    /// Project id or project name.
+    /// Optional project id or project name.
     #[arg(long)]
-    pub project: AzureDevOpsProjectArgument<'static>,
+    pub project: Option<AzureDevOpsProjectArgument<'static>>,
 }
 
 impl AzureDevOpsRepoListArgs {
     pub async fn invoke(self) -> Result<()> {
         let org_url = get_default_organization_url().await?;
 
-        // Find a project matching the provided identifier (id or name).
         let projects = fetch_all_azure_devops_projects(&org_url).await?;
-        let maybe = projects.into_iter().find(|p| self.project.matches(p));
 
-        if let Some(project) = maybe {
-            let repos = fetch_all_azure_devops_repos_for_project(&org_url, &project.id).await?;
+        if let Some(project_filter) = self.project {
+            // Find a project matching the provided identifier (id or name).
+            let maybe = projects.into_iter().find(|p| project_filter.matches(p));
+
+            if let Some(project) = maybe {
+                let repos = fetch_all_azure_devops_repos_for_project(&org_url, &project.id).await?;
+                to_writer_pretty(stdout(), &repos)?;
+                Ok(())
+            } else {
+                eyre::bail!("No project found matching '{}'.", project_filter);
+            }
+        } else {
+            let repo_map = fetch_azure_devops_repos_batch(
+                &org_url,
+                projects.into_iter().map(|p| p.id).collect(),
+            )
+            .await?;
+            let repos = repo_map.into_values().flatten().collect::<Vec<_>>();
             to_writer_pretty(stdout(), &repos)?;
             Ok(())
-        } else {
-            eyre::bail!("No project found matching '{}'.", self.project);
         }
     }
 }