+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add `ct az role operation browse|list` command
 - Add `ct az ad user show {}` command
 - Add `ct az ad oauth2-permission-grant list|create|update|browse` command
+- Update `ct az devops repo list` command to make `--project {}` optional 
 
 # v0.35.1